Refactor let parse to another file:
