+++ src/parser/src/blk.rs
@@ -5,7 +5,7 @@ use crate::Parser;
 use lexer::Token;
 
 impl<'inp> Parser<'inp> {
-    pub fn parse_blk(&mut self, _min_bp: u8) -> Result<Decl, ParseError> {
+    pub(crate) fn parse_blk(&mut self, _min_bp: u8) -> Result<Decl, ParseError> {
         // BlockSeq - vec decls, last expr
         // self.advance(); // put first tok of block into prev_tok
         let blk = self.parse_seq()?;
@@ -13,7 +13,7 @@ impl<'inp> Parser<'inp> {
         let err = format!("Expected '{}' to close block", Token::CloseBrace);
         self.consume_token_type(Token::CloseBrace, &err)?;
 
-        dbg!("prev_tok after blk:", &self.prev_tok);
+        // dbg!("prev_tok after blk:", &self.prev_tok);
         // dbg!("peek after blk:", &self.lexer.peek());
 
         Ok(res)
