+++ src/parser/src/seq.rs
@@ -7,8 +7,9 @@ use lexer::Token;
 use std::rc::Rc;
 
 impl<'inp> Parser<'inp> {
-    pub fn parse_seq(&mut self) -> Result<BlockSeq, ParseError> {
+    pub(crate) fn parse_seq(&mut self) -> Result<BlockSeq, ParseError> {
         let mut decls: Vec<Decl> = vec![];
+        let mut symbols: Vec<String> = vec![];
         let mut last_expr: Option<Expr> = None;
 
         while self.lexer.peek().is_some() {
@@ -21,7 +22,7 @@ impl<'inp> Parser<'inp> {
             // dbg!("prev_tok:", &self.prev_tok);
 
             let expr = self.parse_decl()?;
-            dbg!("Got expr:", &expr);
+            // dbg!("Got expr:", &expr);
             // dbg!("Peek:", &self.lexer.peek());
 
             // end of block: lexer empty OR curly brace (TODO add curly later)
@@ -33,7 +34,13 @@ impl<'inp> Parser<'inp> {
             // let semi = expect_token_body!(Semi, "semicolon");
             // TODO: handle block as expr stmt here - block as last expr was already handled above and we break
             else if self.is_peek_token_type(Token::Semi) {
+                // parse_let doesn't consume the semicolon but does check peek for Semi, so we will definitely run this if expr was let
+                if let Decl::LetStmt(ref stmt) = expr {
+                    symbols.push(stmt.ident.to_owned());
+                }
+
                 decls.push(expr);
+
                 self.advance();
                 // dbg!("Peek after semi:", &self.lexer.peek());
             }
@@ -49,14 +56,6 @@ impl<'inp> Parser<'inp> {
             }
             // Syntax error
             else {
-                dbg!("prev_tok:", &self.prev_tok);
-                let is_brace = &self
-                    .prev_tok
-                    .as_ref()
-                    .map(|tok| tok.eq(&Token::CloseBrace))
-                    .unwrap_or(false);
-                dbg!("prev_tok is close brace:", is_brace);
-                dbg!("peek:", &self.lexer.peek());
                 return Err(ParseError::new("Expected semicolon"));
             }
         }
@@ -64,6 +63,7 @@ impl<'inp> Parser<'inp> {
         Ok(BlockSeq {
             decls,
             last_expr: last_expr.map(Rc::new),
+            symbols,
         })
     }
 }
